moved react to peer deps
This makes it less likely to clash with a user's react version.  If they upgrade to 0.11 when it comes out, your package will automatically use their version.  If you discover it's not compatible with 0.11, you can restrict the peerDependencies to disallow it, and then they'll get a warning/error instead of it silently including both react versions.