[added] Browser build script
There is currently no way to NOT use Browserify's "process" shim
since the code we're requiring in react/lib uses the process object,
so the build is slightly larger than it needs to be. If React core
would expose the stuff in react/lib as modules, we could avoid it.

Fixes #73