Add `location` prop to Route and Switch
As far as we know, this is only useful for animations where you want a subsequent render to use the old props (the way ReactTransitionGroup saves off the previous children).

Let’s hope it doesn’t get abused 🙏🏼

Some notes:

Theoretically, with this change we could ditch the entire subscription model. We could have the top level `Router` subscribe to history and setState, providing location on context. Then Routes would get updates naturally and access location from context. The problem is that `shouldComponentUpdate` will block deep updates to Routes. This is why we introduced subscriptions in the first place. With a location prop, however, apps could pass the location in to Routes that are behind a sCU. For example, in Redux, you would have location in your store, and then make a `ConnectedRoute` that gets the location from there instead of context.

But that’s for a different day. The React team tells us “accessing data on context is unreliable, please use subscriptions” and so we do, which unfortunately opts us completely out of the `setState` update model, which is why this `location` prop is necessary to opt-back into it for animations.

fixes #4351