Publish lock files
After the discussion in afab0726, I wasn't sure the reasons for us not publishing these. It's always a good idea to do so, as it ensures build and development stability as deps update behind the scenes. You can always check what's outdated with npm outdated and update to the latest stuff with npm update.

Feel free to revert if there's any disagreement.